+++ Packages/com.github.kurotu.vrc-quest-tools/Editor/I18n/I18nEnglish.cs
@@ -151,10 +151,10 @@ namespace KRT.VRCQuestTools.I18n
         internal override string ConvertedAvatarEditorMessage => "This component indicates the avatar was converted by VRCQuestTools.";
         internal override string ConvertedAvatarEditorNDMFMessage => "Unsupported components will be removed in NDMF optimization phase.";
 
-        // Platform Component Settings
+        // Platform Component Remover
         internal override string ComponentRequiresNdmf => "This component requires Non-Destructive Modular Framework (NDMF).";
         internal override string BuildTargetLabel => "Build Target";
-        internal override string PlatformComponentSettingsEditorComponentSettingsLabel => "Component Settings";
+        internal override string PlatformComponentRemoverEditorComponentSettingsLabel => "Component Settings";
         internal override string ComponentLabel => "Component";
 
         // NDMF