+++ Packages/com.github.kurotu.vrc-quest-tools/Editor/I18n/I18nJapanese.cs
@@ -150,10 +150,10 @@ namespace KRT.VRCQuestTools.I18n
         internal override string ConvertedAvatarEditorMessage => "このコンポーネントはアバターが VRCQuestTools で変換されたことを示します。";
         internal override string ConvertedAvatarEditorNDMFMessage => "非対応コンポーネントは NDMF の Optimization Phase で削除されます。";
 
-        // Platform Component Settings
+        // Platform Component Remover
         internal override string ComponentRequiresNdmf => "このコンポーネントには Non-Destructive Modular Framework (NDMF) が必要です。";
         internal override string BuildTargetLabel => "ビルドターゲット";
-        internal override string PlatformComponentSettingsEditorComponentSettingsLabel => "コンポーネント設定";
+        internal override string PlatformComponentRemoverEditorComponentSettingsLabel => "コンポーネント設定";
         internal override string ComponentLabel => "コンポーネント";
 
         // NDMF