+++ Packages/com.github.kurotu.vrc-quest-tools/Editor/NDMF/Passes/PlatformComponentRemoverPass.cs
@@ -6,17 +6,17 @@ using UnityEngine;
 namespace KRT.VRCQuestTools.Ndmf
 {
     /// <summary>
-    /// Remove non-active platform components by PlatformComponentSettings in NDMF.
+    /// Remove components by PlatformComponentRemover in NDMF.
     /// </summary>
-    internal class PlatformComponentSettingsPass : Pass<PlatformComponentSettingsPass>
+    internal class PlatformComponentRemoverPass : Pass<PlatformComponentRemoverPass>
     {
         /// <inheritdoc/>
-        public override string DisplayName => "Remove non-active platform components";
+        public override string DisplayName => "Remove platform components";
 
         /// <inheritdoc/>
         protected override void Execute(BuildContext context)
         {
-            var components = context.AvatarRootObject.GetComponentsInChildren<PlatformComponentSettings>(true);
+            var components = context.AvatarRootObject.GetComponentsInChildren<PlatformComponentRemover>(true);
             foreach (var component in components)
             {
                 var buildTarget = component.buildTarget;
@@ -47,12 +47,12 @@ namespace KRT.VRCQuestTools.Ndmf
                         continue;
                     }
 
-                    if (buildTarget == Models.BuildTarget.PC && !settings.enabledOnPC)
+                    if (buildTarget == Models.BuildTarget.PC && settings.removeOnPC)
                     {
                         Debug.Log($"Remove {settings.component.GetType().Name} for PC platform", settings.component.gameObject);
                         Object.DestroyImmediate(settings.component);
                     }
-                    if (buildTarget == Models.BuildTarget.Android && !settings.enabledOnAndroid)
+                    if (buildTarget == Models.BuildTarget.Android && settings.removeOnAndroid)
                     {
                         Debug.Log($"Remove {settings.component.GetType().Name} for Android platform", settings.component.gameObject);
                         Object.DestroyImmediate(settings.component);